// front-gention-bitacoras/src/app/model/request/consultores.request.tsx
export class GuardaConsultor {
    id_consultor: number | null;
    nombre: string;
    primerAp: string;
    segundoAp: string;
    correo: string;
    correoSecundario: string;
    curp: string;
    rfc: string;
    id_archivo: number | null;
    archivoCurriculum: string;
    archivoCertificado: string;
    archivoOtros: string;
  }  